+++ packages/deprecation-crawler/src/lib/commands/group.command.ts
@@ -1,4 +1,4 @@
-import { getConfig } from '../config';
+import { setup } from '../processors/setup';
 import { addGroups } from '../tasks/add-groups';
 import { generateOutput } from '../tasks/generate-output';
 import { CrawlConfig, CrawledRelease, CrawlerProcess } from '../models';
@@ -11,7 +11,7 @@ export const groupCommand: YargsCommandObject = {
   module: {
     handler: async (argv) => {
       if (argv.verbose) console.info(`run grouping as a yargs command`);
-      const config = await getConfig();
+      const config = await setup();
       const tasks = [loadExistingDeprecations, addGroups, generateOutput];
 
       // Run all processors
