+++ packages/deprecation-crawler/src/lib/constants.ts
@@ -1,3 +1,8 @@
+import { Options } from 'yargs';
+import { YargsCommandObject } from './cli/model';
+import { groupCommand } from './commands/group.command';
+import { defaultCommand } from './commands/default.command';
+
 export const DEPRECATIONS_OUTPUT_DIRECTORY = 'deprecations';
 export const CRAWLER_CONFIG_PATH = 'deprecation-crawler.config.json';
 export const SEMVER_TOKEN = `SEMVER_TOKEN`;
@@ -17,3 +22,28 @@ export const enum CRAWLER_MODES {
 }
 export const MD_GROUP_OPENER = '<!-- ruid-groups';
 export const MD_GROUP_CLOSER = 'ruid-groups -->';
+
+export const OPTIONS: { [key: string]: Options } = {
+  verbose: {
+    alias: 'v',
+    type: 'boolean',
+    description: 'Run with verbose logging',
+  },
+  path: {
+    alias: 'p',
+    type: 'string',
+    description: 'Path to deprecation-crawler.config.json',
+  },
+  tag: {
+    alias: 't',
+    type: 'string',
+    description: 'Tag to crawler',
+  },
+  'next-version': {
+    alias: 'n',
+    type: 'string',
+    description: 'Version of the crawled results',
+  },
+};
+
+export const COMMANDS: YargsCommandObject[] = [groupCommand, defaultCommand];
