+++ packages/deprecation-crawler/src/lib/index.ts
@@ -1,7 +1,7 @@
 import { isCrawlerModeCi } from './utils';
 import { guardAgainstDirtyRepo } from './tasks/guard-against-dirty-repository';
 import { runCli } from './cli';
-import { COMMANDS, OPTIONS } from './cli-settings';
+import { COMMANDS, OPTIONS } from './constants';
 
 (async () => {
   if (isCrawlerModeCi()) {
