+++ packages/deprecation-crawler/src/lib/output-formatters/index.ts
@@ -1,12 +1,11 @@
+import { generateTagBasedFormat } from './markdown/tag-based.md-formatter';
+import { generateGroupBasedFormat } from './markdown/group-based.md.formatter';
 import { generateDeprecationIndex } from './markdown/deprecation-index.md.formatter';
-export { addCommentToRepository } from './git/tag-comments.git.formatter';
-export {generateRawJson} from './json/raw.json.formatter';
-import {generateTagBasedFormatter} from './markdown/tag-based.md-formatter';
-import {generateGroupBasedFormatter} from './markdown/group-based.md.formatter';
-
+export { generateTaggedCommentsInRepository } from './git/tag-comments.git.formatter';
+export { generateRawJson } from './json/raw.json.formatter';
 
 export const builtInFormatter = {
-  tagBasedMarkdown: generateTagBasedFormatter,
-  groupBasedMarkdown: generateGroupBasedFormatter,
+  tagBasedMarkdown: generateTagBasedFormat,
+  groupBasedMarkdown: generateGroupBasedFormat,
   deprecationIndex: generateDeprecationIndex,
 };
